Move Table and Layout extension to file relational.rs
